+++ lib/generators/web_angular_data.dart
@@ -1,0 +1,46 @@
+// Copyright (c) 2014, Google Inc. Please see the AUTHORS file for details.
+// All rights reserved. Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+const List<String> data = const [
+  ".gitignore",
+  "text",
+  """IyBGaWxlcyBhbmQgZGlyZWN0b3JpZXMgY3JlYXRlZCBieSBwdWIKLmJ1aWxkbG9nCi5wYWNrYWdl
+cwoucHJvamVjdAoucHViLwpidWlsZC8KcGFja2FnZXMKcHVic3BlYy5sb2NrICMgKFJlbW92ZSB0
+aGlzIHBhdHRlcm4gaWYgeW91IHdpc2ggdG8gY2hlY2sgaW4geW91ciBsb2NrIGZpbGUpCgojIEZp
+bGVzIGNyZWF0ZWQgYnkgZGFydDJqcwoqLmRhcnQuanMKKi5wYXJ0LmpzCiouanMuZGVwcwoqLmpz
+Lm1hcAoqLmluZm8uanNvbgoKIyBEaXJlY3RvcnkgY3JlYXRlZCBieSBkYXJ0ZG9jCmRvYy9hcGkv
+CgojIEpldEJyYWlucyBJREVzIChXZWJTdG9ybSBhbmQgSURFQSBhcmUgdGhlIHJlY29tZW5kZWQg
+RGFydCBJREVzKSAKLmlkZWEvCiouaW1sCiouaXByCiouaXdzCg==""",
+  "lib/app_component.dart",
+  "text",
+  """Ly8gQ29weXJpZ2h0IChjKSBfX3llYXJfXywgX19hdXRob3JfXy4gQWxsIHJpZ2h0cyByZXNlcnZl
+ZC4gVXNlIG9mIHRoaXMgc291cmNlIGNvZGUKLy8gaXMgZ292ZXJuZWQgYnkgYSBCU0Qtc3R5bGUg
+bGljZW5zZSB0aGF0IGNhbiBiZSBmb3VuZCBpbiB0aGUgTElDRU5TRSBmaWxlLgoKbGlicmFyeSBf
+X3Byb2plY3ROYW1lX18uYXBwX2NvbXBvbmVudDsKCmltcG9ydCAncGFja2FnZTphbmd1bGFyMi9h
+bmd1bGFyMi5kYXJ0JzsKCkBDb21wb25lbnQoc2VsZWN0b3I6ICdteS1hcHAnLCB0ZW1wbGF0ZVVy
+bDogJ2FwcF9jb21wb25lbnQuaHRtbCcpCmNsYXNzIEFwcENvbXBvbmVudCB7fQo=""",
+  "lib/app_component.html",
+  "text",
+  "PGgxPk15IEZpcnN0IEFuZ3VsYXIgMiBBcHA8L2gxPgo=",
+  "pubspec.yaml",
+  "text",
+  """bmFtZTogX19wcm9qZWN0TmFtZV9fCnZlcnNpb246IDAuMC4xCmRlcGVuZGVuY2llczoKICBhbmd1
+bGFyMjogMi4wLjAtYWxwaGEuNDcKICBicm93c2VyOiBeMC4xMC4wCnRyYW5zZm9ybWVyczoKLSBh
+bmd1bGFyMjoKICAgIGVudHJ5X3BvaW50czogd2ViL21haW4uZGFydAo=""",
+  "web/index.html",
+  "text",
+  """PCFET0NUWVBFIGh0bWw+CjxodG1sPgogIDxoZWFkPgogICAgPHRpdGxlPl9fcHJvamVjdE5hbWVf
+XzwvdGl0bGU+CgogICAgPHNjcmlwdCBkZWZlciBzcmM9Im1haW4uZGFydCIgdHlwZT0iYXBwbGlj
+YXRpb24vZGFydCI+PC9zY3JpcHQ+CiAgICA8c2NyaXB0IGRlZmVyIHNyYz0icGFja2FnZXMvYnJv
+d3Nlci9kYXJ0LmpzIj48L3NjcmlwdD4KICA8L2hlYWQ+CiAgPGJvZHk+CiAgICA8bXktYXBwPkxv
+YWRpbmcuLi48L215LWFwcD4KICA8L2JvZHk+CjwvaHRtbD4K""",
+  "web/main.dart",
+  "text",
+  """Ly8gQ29weXJpZ2h0IChjKSBfX3llYXJfXywgX19hdXRob3JfXy4gQWxsIHJpZ2h0cyByZXNlcnZl
+ZC4gVXNlIG9mIHRoaXMgc291cmNlIGNvZGUKLy8gaXMgZ292ZXJuZWQgYnkgYSBCU0Qtc3R5bGUg
+bGljZW5zZSB0aGF0IGNhbiBiZSBmb3VuZCBpbiB0aGUgTElDRU5TRSBmaWxlLgoKbGlicmFyeSBf
+X3Byb2plY3ROYW1lX18ud2ViOwoKaW1wb3J0ICdwYWNrYWdlOmFuZ3VsYXIyL2Jvb3RzdHJhcC5k
+YXJ0JzsKCmltcG9ydCAncGFja2FnZTpfX3Byb2plY3ROYW1lX18vYXBwX2NvbXBvbmVudC5kYXJ0
+JzsKCm1haW4oKSB7CiAgYm9vdHN0cmFwKEFwcENvbXBvbmVudCk7Cn0K"""
+];