+++ lib/stagehand.dart
@@ -31,6 +31,7 @@ import 'generators/console_simple.dart';
 import 'generators/package_simple.dart';
 import 'generators/server_appengine.dart';
 import 'generators/server_shelf.dart';
+import 'generators/web_angular.dart';
 import 'generators/web_polymer.dart';
 import 'generators/web_simple.dart';
 import 'src/common.dart';
@@ -42,6 +43,7 @@ final List<Generator> generators = [
   new PackageSimpleGenerator(),
   new ServerAppEngineGenerator(),
   new ServerShelfGenerator(),
+  new WebAngularGenerator(),
   new WebPolymerGenerator(),
   new WebSimpleGenerator()
 ]..sort();