+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.4
 - Migrated `server-appengine` to use `shelf`.
+- Updated the template `.gitignore` files
 
 ## 0.2.3+1
 - Fixed an issue with the Polymer template